MAJOR CHANGES: Refactored susceptibility functions
Now using Numba to optimise the population of the forward matrix in the
multipole inversion code.
Quick tests show a speed up of 30x for the matrix generation, helping to
the calculation of multiple tests. The first-time the sus functions are
called need to be compiled thus speed up is noticed from a second call
onwards.
For now only the spherical_harmonics_basis.py module has been optimised.
The functions accept all the scan surface positions at once and populate
the Q matrix in one single loop. Strides are necessary to populate the
matrix according to the user specified multipole expansion order.
Errors seem to slightly change but the outputs remain basically the
same.
The test notebook is currently in the mmt_scripts repository and
will be added here later.